+++ packages/editor/src/components/collab-sidebar/hooks.js
@@ -270,7 +270,7 @@ function clearInlineNoteMarker(
 export function useNoteActions() {
 	const { createNotice } = useDispatch( noticesStore );
 	const { saveEntityRecord, deleteEntityRecord } = useDispatch( coreStore );
-	const { getCurrentPostId } = useSelect( editorStore );
+	const { getNotesPostId } = unlock( useSelect( editorStore ) );
 	const {
 		getBlockAttributes,
 		getClientIdsWithDescendants,
@@ -307,7 +307,7 @@ export function useNoteActions() {
 				'root',
 				'comment',
 				{
-					post: getCurrentPostId(),
+					post: getNotesPostId(),
 					content,
 					status: 'hold',
 					type: 'note',
@@ -385,7 +385,7 @@ export function useNoteActions() {
 
 				// Then create a new note with the metadata.
 				const newNoteData = {
-					post: getCurrentPostId(),
+					post: getNotesPostId(),
 					content: content || '', // Empty content for resolve, content for reopen.
 					type: 'note',
 					status,